refactor Playlist and Song into /data
