+++ app/src/androidTest/java/com/github/fribourgsdp/radio/ImportSpotifyPlaylistsActivityTest.kt
@@ -1,5 +1,6 @@
 package com.github.fribourgsdp.radio
 
+import com.github.fribourgsdp.radio.data.Song
 import junit.framework.TestCase.assertEquals
 import okhttp3.*
 import org.json.JSONArray
