+++ app/src/androidTest/java/com/github/fribourgsdp/radio/PlaylistTest.kt
@@ -1,5 +1,7 @@
 package com.github.fribourgsdp.radio
 
+import com.github.fribourgsdp.radio.data.Playlist
+import com.github.fribourgsdp.radio.data.Song
 import com.google.android.gms.tasks.Tasks
 import org.junit.Test
