+++ app/src/androidTest/java/com/github/fribourgsdp/radio/UserTest.kt
@@ -3,6 +3,7 @@ package com.github.fribourgsdp.radio
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.fribourgsdp.radio.data.Playlist
 import org.junit.Assert.*
 import org.junit.Test
 import org.junit.runner.RunWith
