+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/MockGameSettingsActivity.kt
@@ -3,6 +3,8 @@ package com.github.fribourgsdp.radio.mockimplementations
 import android.content.Context
 import android.os.Bundle
 import com.github.fribourgsdp.radio.*
+import com.github.fribourgsdp.radio.data.Playlist
+import com.github.fribourgsdp.radio.data.Song
 import org.mockito.Mockito.*
 
 class MockGameSettingsActivity: GameSettingsActivity() {
