+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/MockLobbyActivities.kt
@@ -2,12 +2,9 @@ package com.github.fribourgsdp.radio.mockimplementations
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
-import androidx.core.content.ContextCompat.startActivity
 import com.github.fribourgsdp.radio.*
-import com.github.fribourgsdp.radio.mockimplementations.BuggyDatabase
-import com.github.fribourgsdp.radio.mockimplementations.LocalDatabase
-import kotlinx.serialization.decodeFromString
+import com.github.fribourgsdp.radio.data.Playlist
+import com.github.fribourgsdp.radio.data.Song
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
