+++ app/src/main/java/com/github/fribourgsdp/radio/PlaylistSongsFragment.kt
@@ -1,17 +1,14 @@
 package com.github.fribourgsdp.radio
 
-import android.content.ContentValues
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import android.view.View
 import android.widget.Button
 import android.widget.TextView
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.google.android.gms.tasks.Task
-import com.google.android.gms.tasks.Tasks
-import kotlinx.serialization.decodeFromString
+import com.github.fribourgsdp.radio.data.Playlist
+import com.github.fribourgsdp.radio.data.Song
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
