+++ app/src/main/java/com/github/fribourgsdp/radio/User.kt
@@ -2,6 +2,8 @@ package com.github.fribourgsdp.radio
 
 
 import android.content.Context
+import com.github.fribourgsdp.radio.data.Playlist
+import com.github.fribourgsdp.radio.data.Song
 
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
