+++ app/src/main/java/com/github/fribourgsdp/radio/data/Nameable.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.data
 
 interface Nameable {
     var name: String
